+++ lua/obsidian/init.lua
@@ -102,6 +102,10 @@ obsidian.setup = function(opts)
     require("obsidian.statusline").start(client)
   end
 
+  if opts.footer.enabled then
+    require("obsidian.footer").start(client)
+  end
+
   -- Register completion sources, providers
   if opts.completion.nvim_cmp then
     require("obsidian.completion.plugin_initializers.nvim_cmp").register_sources(opts)