+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New `obsidian.config` type for user config type check.
 - More informative healthcheck.
 - A guide to embed images for both viewing in neovim and obsidian app: https://github.com/obsidian-nvim/obsidian.nvim/wiki/Images
+- Added footer options.
 
 ### Changed
 
@@ -47,6 +48,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use a `Obsidian` global variable to hold the state instead of client.
 - `opts.img_text_func` has an obsidian app compatibility, and only accept one path argument.
 - Moved `client:apply_async` -> `api.dir`
+- Deprecate `statusline` options and `vim.g.obsidian` in favor of `footer` options and `vim.b.obsidian_footer`.
 
 ### Fixed